requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-SocketIO==5.3.4
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
lxml==4.9.3
MarkupSafe==2.1.2
mouse==0.7.1
MouseInfo==0.1.3
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pefile==2023.2.7
Pillow==9.5.0
psutil==5.9.5
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyinstaller-versionfile==2.1.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==6.0
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.6
zope.interface==6.0
